This file uses all 3 buckling loads and correctly recomputes Iz
This file checks all 3 buckling loads, full shear, full compress and combined, depending on the position on the fuselage.

It also correctly recomputes Iz.

Currently unforunately it is producing rather thicc layups. I'm not sure if it's a minor error in a formula but it seems like everything is implemented correctly?

I believe this contains everything that could possibly cause the laminate to fail though. So even if it's a 222% heavier than aluminium. It's a safe one.